Adds AddTransaction component
